# openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery Driver App API
  version: 1.6.1
  description: >
    API for driver app, admin dashboard, customer tracking, notifications,
    subscription deliveries, wallet & payout management, and route optimization.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: driverId
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags: [Orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum:
                    - created
                    - assigned
                    - picked_up
                    - out_for_delivery
                    - delivered
                    - cancelled
                cancellationSource:
                  type: string
                  enum:
                    - customer
                    - restaurant
                    - admin
                    - system
                  nullable: true
                cancellationReason:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        externalRef:
          type: string
          nullable: true
        pickup:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            address:
              type: string
        dropoff:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            address:
              type: string
        status:
          type: string
          enum:
            - created
            - assigned
            - picked_up
            - out_for_delivery
            - delivered
            - cancelled
        cancellationSource:
          type: string
          enum:
            - customer
            - restaurant
            - admin
            - system
          nullable: true
        cancellationReason:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: string
        paymentType:
          type: string
          enum:
            - cod
            - prepaid
            - wallet
        slaSeconds:
          type: integer
          nullable: true
        trackingUrl:
          type: string
          nullable: true
        zoneId:
          type: string
          nullable: true
        subscriptionId:
          type: string
          nullable: true
        assignedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
    OrderInput:
      allOf:
        - $ref: '#/components/schemas/Order'
      required:
        - pickup
        - dropoff
        - paymentType

    # Proof of Delivery
...